Matrix direction
